requirements.txt
garminconnect
pandas
matplotlib
python-dotenv 